Dropping "No derivative works" from license (still requiring non-commercial use and attribution). An example of a fork I don't mind is here:
http://userscripts.org/scripts/show/69740

He clearly gives attribution, and repurposes the script to be, well, light. No gui and such, as it was back in it's older form. He clearly did some work cutting the script down, instead of just copying and pasting my script, which would've just been pollution.

Removed comments at the beginning of the script, the note about editing refers to very old versions of this script and don't apply anymore.

I believe current users of the script are familiar with requesting features, and that future users will be smart enough to figure out the procedure on their own.